Moving custom components to mod file to avoid circular dependencies/lazy loading issues (#2580)
