+++ src/custom/components/swap/SwapModalHeader/SwapModalHeaderMod.tsx
@@ -28,7 +28,7 @@ import { AMOUNT_PRECISION, INPUT_OUTPUT_EXPLANATION } from 'constants/index'
 import { computeSlippageAdjustedAmounts } from 'utils/prices'
 import { Field } from 'state/swap/actions'
 import { formatMax, formatSmart } from 'utils/format'
-import { AuxInformationContainer } from 'components/CurrencyInputPanel'
+import { AuxInformationContainer } from 'components/CurrencyInputPanel/CurrencyInputPanelMod'
 import FeeInformationTooltip from '../FeeInformationTooltip'
 import { LightCardType } from '.'
 import { transparentize } from 'polished'
